requirements.txt
django-qsstats-magic
python-dateutil
